requirements.txt
transformers
torch
pytorchvideo
ffmpeg-python
torchvision
tqdm
fathomnet
openai
wandb
pafy
boto3
